refactor: extract shared server bootstrap helpers
